# openapi.yaml
openapi: 3.0.3
info:
  title: Tienda de Videojuegos DEVCO
  description: |-
    API para una tienda de videojuegos, que permite realizar las operaciones
    básicas (CRUD). 
    
    Permite gestionar los videojuegos, las ordenes y los
    usuarios de la tienda.
  contact:
    email: asalazars@devco.com.co
  version: 1.0.0
...
paths:
  /videojuego:
    get:
      tags:
        - videojuego
      summary: Obtiene todos los videojuegos de la tienda
      description: Obtiene todo el inventario de videojuegos de la tienda
      operationId: obtenerTodosVideojugos
      responses:
        '200':
          description: Operación exitosa
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Videojuego'
    post:
      tags:
        - videojuego
      summary: Agrega un nuevo videojuego a la tienda
      description: Agrega un nuevo videojuego al inventario de la tienda
      operationId: crearVideojuego
      requestBody:
        description: Crea un nuevo videojuego en la tienda
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                  minLength: 1
                  example: videojuego
                precio:
                  type: number
                  example: 100000
                imagen:
                  type: string
                  minLength: 1
                  example: http://url-image.com
                stock:
                  type: integer
                  example: 5
          application/xml:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                  minLength: 1
                  example: videojuego
                precio:
                  type: number
                  example: 100000
                imagen:
                  type: string
                  minLength: 1
                  example: http://url-image.com
                stock:
                  type: integer
                  example: 5
              xml:
                name: videojuego
        required: true
      responses:
        '201':
          description: Creación exitosa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Videojuego'          
            application/xml:
              schema:
                $ref: '#/components/schemas/Videojuego'
        '400':
          description: Datos inválidos
      security:
        - bearerAuth: []
  /videojuego/{IdVideojuego}:
    get:
      tags:
        - videojuego
      summary: Obtiene un videojuego por su id
      description: Devuelve un único videojuego
      operationId: obtenerVideojuegoPorId
      parameters:
        - name: IdVideojuego
          in: path
          description: ID del videojuego a obtener
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Operación exitosa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Videojuego'          
            application/xml:
              schema:
                $ref: '#/components/schemas/Videojuego'
        '400':
          description: ID inválido
        '404':
          description: Videojuego no encontrado
    put:
      tags:
        - videojuego
      summary: Actualiza un videojuego
      description: Actualiza un videojuego con los datos suministrados
      operationId: actualizarVideojuego
      parameters:
        - name: IdVideojuego
          in: path
          description: ID del videojuego a actualizar
          required: true
          schema:
            type: integer
      requestBody:
        description: Actualiza un videojuego de la tienda
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                  minLength: 1
                  example: videojuego
                precio:
                  type: number
                  example: 100000
                imagen:
                  type: string
                  minLength: 1
                  example: http://url-image.com
                stock:
                  type: integer
                  example: 5
          application/xml:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                  minLength: 1
                  example: videojuego
                precio:
                  type: number
                  example: 100000
                imagen:
                  type: string
                  minLength: 1
                  example: http://url-image.com
                stock:
                  type: integer
                  example: 5
              xml:
                name: videojuego
        required: true
      responses:
        '201':
          description: Actualización exitosa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Videojuego'          
            application/xml:
              schema:
                $ref: '#/components/schemas/Videojuego'
        '400':
          description: Datos inválidos
      security:
        - bearerAuth: []
    delete:
      tags:
        - videojuego
      summary: Elimina un videojuego
      description: Elimina un videojuego del inventario
      operationId: eliminarVideojuego
      parameters:
        - name: IdVideojuego
          in: path
          description: ID del videojuego a actualizar
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Operación exitosa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Videojuego'          
            application/xml:
              schema:
                $ref: '#/components/schemas/Videojuego'
        '400':
          description: ID inválido
        '404':
          description: Videojuego no encontrado
      security:
        - bearerAuth: []
...
components:
  schemas:
    Videojuego:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nombre:
          type: string
          minLength: 1
          example: videojuego
        precio:
          type: number
          example: 100000
        imagen:
          type: string
          minLength: 1
          example: http://url-image.com
        stock:
          type: integer
          example: 5
      xml:
        name: videojuego
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT